+++ src/main.rs
@@ -26,6 +26,7 @@ mod lib {
     pub mod favourites;
     pub mod ghratask;
     pub mod ghrtask;
+    pub mod goget;
     pub mod python;
     pub mod rust;
     pub mod ssh;